test/requirements.txt
mcp>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pyyaml>=6.0.0